+++ sn_node/CHANGELOG.md
@@ -2924,6 +2924,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - *(node)* change RPC log level to debug
 - *(test)* connect to RPC endpoint with retries
 
+## [0.103.5](https://github.com/maidsafe/safe_network/compare/sn_node-v0.103.4...sn_node-v0.103.5) - 2024-01-18
+
+### Other
+- update dependencies
+
+## [0.103.6](https://github.com/maidsafe/safe_network/compare/sn_node-v0.103.5...sn_node-v0.103.6) - 2024-01-21
+
+### Other
+- update dependencies
+
+## [0.103.7](https://github.com/maidsafe/safe_network/compare/sn_node-v0.103.6...sn_node-v0.103.7) - 2024-01-22
+
+### Other
+- update dependencies
+
 ## v0.1.0 (2023-05-04)
 
 ### Chore