+++ sn_testnet/CHANGELOG.md
@@ -2577,6 +2577,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Other
 - update dependencies
 
+## [0.3.26](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.3.25...sn_testnet-v0.3.26) - 2024-01-18
+
+### Other
+- update dependencies
+
+## [0.3.27](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.3.26...sn_testnet-v0.3.27) - 2024-01-21
+
+### Other
+- update dependencies
+
+## [0.3.28](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.3.27...sn_testnet-v0.3.28) - 2024-01-22
+
+### Other
+- update dependencies
+
 ## v0.1.0 (2023-03-16)
 
 <csr-id-4f04bd1a5d1c747bfc6b5d39824dd108f8546b7b/>